Create README.md
For the project I created a web service that allows a variety ways of interaction using industry standard methods (AJAX and jQuery), which generates and utilizes different data formats (Text, XML and JSON). In the DAO I work through a MySQL database that has a table called 'books' for CRUD operations.